+++ CHANGELOG.md
@@ -45,5 +45,6 @@
 ## 3.0.0
 
 - Add support for Node.js v13 by no longer extending `ReadStream` and `WriteStream` from node's `fs` library.
+- Specify `0o600` for buffer file permissions instead of node's default `0o666`
 - **BREAKING:** Remove several undocumented properties that existed for consistency with the extended classes.
 - **BREAKING:** No longer listen for `SIGINT`, and instead warn the application to add handlers for terminating signals.